# Conditionally allocate WaitGroup memory